simplifed ceci hotfix implementation
